refactor: Move header and extract back button
