+++ apps/web/src/components/layout/components/LocalStudioHeader/LocalStudioHeader.tsx
@@ -1,24 +1,26 @@
 import { Header } from '@mantine/core';
-import { IconButton, Text } from '@novu/novui';
+import { IconButton } from '@novu/novui';
 import { css } from '@novu/novui/css';
-import { IconHelpOutline, IconOutlineArrowBack } from '@novu/novui/icons';
+import { IconHelpOutline } from '@novu/novui/icons';
 import { HStack } from '@novu/novui/jsx';
-import { hstack } from '@novu/novui/patterns';
 import { FC } from 'react';
-import { discordInviteUrl } from '../../../pages/quick-start/consts';
-import { DocsButton } from '../../docs/DocsButton';
-import { useLocation, useNavigate } from 'react-router-dom';
-import { ROUTES } from '../../../constants/routes';
-import { HEADER_NAV_HEIGHT } from '../constants';
-import { BridgeMenuItems } from './v2/BridgeMenuItems';
+import { matchPath, useLocation, useParams } from 'react-router-dom';
+import { ROUTES } from '../../../../constants/routes';
+import { discordInviteUrl } from '../../../../pages/quick-start/consts';
+import { useStudioNavigate } from '../../../../studio/hooks/useStudioNavigate';
+import { DocsButton } from '../../../docs/DocsButton';
+import { HEADER_NAV_HEIGHT } from '../../constants';
+import { BridgeMenuItems } from '../v2/BridgeMenuItems';
+import { BackButton } from './BackButton';
+
+const HOME_ROUTE: ROUTES = ROUTES.STUDIO_FLOWS_VIEW;
 
 export const LocalStudioHeader: FC = () => {
-  const navigate = useNavigate();
   const { pathname } = useLocation();
+  const navigate = useStudioNavigate();
+  const { templateId = '' } = useParams();
 
-  if (pathname.startsWith(ROUTES.STUDIO_ONBOARDING)) {
-    return null;
-  }
+  const shouldHideBackButton = matchPath(HOME_ROUTE, pathname);
 
   return (
     <Header
@@ -34,21 +36,7 @@ export const LocalStudioHeader: FC = () => {
       <HStack justifyContent="space-between" width="full" display="flex">
         <HStack gap="100">
           {/** TODO temporary back-button. To be refined later */}
-          <button
-            className={hstack({
-              cursor: 'pointer',
-              gap: 'margins.icons.Icon20-txt',
-              px: '75',
-              py: '25',
-              _hover: { opacity: 'hover' },
-            })}
-            onClick={() => navigate(-1)}
-          >
-            <IconOutlineArrowBack />
-            <Text fontWeight="strong" color="typography.text.secondary">
-              Back
-            </Text>
-          </button>
+          {!shouldHideBackButton && <BackButton onClick={() => navigate(HOME_ROUTE, { templateId })} />}
         </HStack>
         <HStack gap="100">
           <BridgeMenuItems />
