+++ apps/web/src/components/layout/components/LocalStudioHeader/index.ts
@@ -1,0 +1,1 @@
+export * from './LocalStudioHeader';
