+++ apps/web/src/components/layout/components/LocalStudioSidebar/LocalStudioSidebarContent.tsx
@@ -34,13 +34,9 @@ export const LocalStudioSidebarContent: FC<LocalStudioSidebarContentProps> = ({ 
         <DocsButton
           tooltip={'Open a guide'}
           TriggerButton={({ onClick }) => (
-            <button onClick={onClick}>
+            <button onClick={onClick} className={css({ width: 'full' })}>
               <div
-                className={cx(
-                  hstack({ cursor: 'pointer' }),
-                  css({ justifyContent: 'flex-start' }),
-                  css(rawButtonBaseStyles)
-                )}
+                className={cx(hstack({ cursor: 'pointer', justifyContent: 'flex-start' }), css(rawButtonBaseStyles))}
               >
                 <IconAdd />
                 <span>Add a workflow</span>
